% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plot.R
\name{avg.plot}
\alias{avg.plot}
\title{Average values plot function}
\usage{
avg.plot(
  dat,
  iter,
  x.lab = "Iteration number",
  y.lab = "Average of simulated value",
  title = NULL,
  details = FALSE
)
}
\arguments{
\item{dat}{Dataset containing the simulated values}

\item{iter}{Iteration number of the simulation}

\item{x.lab}{Label of the x axis in the generated plot, default is set to be "Iteration number"}

\item{y.lab}{Label of the y axis in the generated plot, default is set to be "Average of simulated values"}

\item{title}{Title of the generated plot}

\item{details}{Logical variable to specify whether the average simulate values are returned, default is set to TRUE}
}
\description{
This function generates average simulated values of the parameter values
}
