% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{calcu.acf}
\alias{calcu.acf}
\title{Autocorrelation function}
\usage{
calcu.acf(data, lag = 50, plot = TRUE, title = NULL, details = FALSE)
}
\arguments{
\item{data}{dataset containing the variables of which autocorrelations are calculated.}

\item{lag}{lag at which the autocorrelation is calculated, default is set as 50.}

\item{plot}{logical variable to specify whether the plot is generated, default is set to TRUE.}

\item{title}{title of the generated autocorrelation plots.}

\item{details}{boolean variable to specify whether the autocorrelation values are returned, default is set to FALSE.}
}
\value{
A matrix containing the calculated autocorrelations of all the variables in the dataset. If \code{plot} = TRUE,
the autocorrelation plots of all the variables will be drawn.
}
\description{
Calculates the autocorrelations
}
\details{
This function calculates the autocorrelation values of all the variables on a column by column base.
The default value of \code{lag} is set as 50, the maximum number of lag should not exceed the number of row of the dataset,
which reflects the corresponding number of iteration of running the multiple imputation.
}
