% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calcu.param}
\alias{calcu.param}
\title{CC and AC parameters calculation function}
\usage{
calcu.param(dat)
}
\arguments{
\item{dat}{dataset containing the variables with missing and censored values.}
}
\value{
a list of length 5 containing the avaliable-case mean and variance for each variable, the complete-case covariance matrix,
the complete-case mean and variance for each variable.
}
\description{
Calculates the complete-case and avaliable-case parameters
}
