% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{conv.plot}
\alias{conv.plot}
\title{Convergence plot function}
\usage{
conv.plot(
  data,
  iter,
  x.lab = "Iteration number",
  y.lab = "Simulated values",
  title = NULL
)
}
\arguments{
\item{data}{dataset containing the simulated values}

\item{iter}{number of iterations for running multiple imputation}

\item{x.lab}{label of the x axis in the generated plot, default is set to "Iteration number"}

\item{y.lab}{label of the y axis in the generated plot, default is set to "Simulated values"}

\item{title}{title in the convergence plot.}
}
\value{
The plot of simulated values across iterations.
}
\description{
Generates the convergence plots for the parameter values
}
\details{
The function generates the trace plot of simulated values across iterations.
}
