% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datagen.R
\name{dat.gen}
\alias{dat.gen}
\title{Data generation function}
\usage{
dat.gen(n, m.vec, sig, p, miss.num, censor.num)
}
\arguments{
\item{n}{Number of observations to be generated}

\item{m.vec}{Specified mean vector}

\item{sig}{Specified variance-covariance matrix}

\item{p}{Number of variables to be generated}

\item{miss.num}{Number of variables having missing values}

\item{censor.num}{NUmber of variables having censored values}
}
\description{
Users can specify the total number of variables in total, the number of variables that have missing values and the number of variables that have censored values
}
