% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.generation.R
\name{data.generation}
\alias{data.generation}
\title{Data generating function}
\usage{
data.generation(
  n,
  p,
  mu,
  Sigma,
  miss.pos,
  miss.percent,
  miss.type = "MCAR",
  censor.pos,
  censor.percent1,
  censor.percent2,
  censor.type = "interval"
)
}
\arguments{
\item{n}{number of observations to be generated}

\item{p}{number of variables to be generated}

\item{mu}{specified mean vector}

\item{Sigma}{specified covariance matrix}

\item{miss.pos}{indexes of the variables that have missing values}

\item{miss.percent}{the missing percentage in each variable that has missing values}

\item{miss.type}{mechanism of missing data, available options include the default \code{"MCAR"} for missing completely at random, and
\code{"MAR"} for missing at random}

\item{censor.pos}{indexes of the variables that have censored values}

\item{censor.percent1}{the percentage of the lower limit of the censored values}

\item{censor.percent2}{the percentage of the upper limit of the censored values}

\item{censor.type}{type of censored data, available options include the default \code{"interval"} for interval censoring, \code{"right"}
for right censoring, and \code{"left"} for left censoring}
}
\value{
A list of two containing two matrices.
}
\description{
Generates multivariate normal data with missing and censored values
}
\details{
This function generates the multivariate normal data  with missing and censored values, that can be used to verify the
correctness of the multiple imputation algorithm. Users have to specify the sample size \code{n}, the dimension \code{p}, the
mean vector \code{mu}, and the covariance matrix \code{Sigma} of the data. Additionally, the information for the missing and censoring,
such as the indexes, percentages for missing and censored values, missing mechanisms or censoring types, should also be included.
Specifically, we assign a pair of values, say \code{X_{ll}} and \code{X_{ul}}, to each of the data point in the following way: if the
values is observed, we set \code{X_{ll}} = \code{X_{ul}}; if the value is missing, we set \code{X_{ll}} = -10e10 and \code{X_{ul}} =
10e10; if the value is censored, we set \code{X_{ll}} and \code{X_{ll}} to finite values according to the specified censoring type.
}
