% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{marg.plot}
\alias{marg.plot}
\title{Marginal plots function}
\usage{
marg.plot(dat, title = NULL)
}
\arguments{
\item{dat}{Dataset containing each variable}

\item{title}{Titles of the generated plots}
}
\description{
This function draws the marginal density plots for each variable
}
