% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_gen.R
\name{multi.dat.gen}
\alias{multi.dat.gen}
\title{Data generation function}
\usage{
multi.dat.gen(n, m.vec, sig, p, miss.num, censor.num)
}
\arguments{
\item{n}{Number of observations to be generated}

\item{m.vec}{Specified mean vector}

\item{sig}{Specified variance-covariance matrix}

\item{p}{Number of variables to be generated}

\item{miss.num}{Number of variables having missing values}

\item{censor.num}{NUmber of variables having censored values}
}
\value{
A list containing the full data, the complete data with missing and censoring information applied,
  the missing and censoring index, positions of the variables including the missing or censored values, or
  the fully observed variable, and the two limits of the censored values
}
\description{
Function to generate multivariate data. Users can specify the total number of variables in total, the number of variables that have missing values and the number of variables that have censored values
}
\examples{
n <- 1000
p <- 3
m.vec <- c(3, 2, 1)
V <- diag(p)
miss.num <- 1
censor.num <- 1
dat <- multi.dat.gen(n, m.vec, V, p, miss.num, censor.num)

}
