% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnimpute-package.R
\docType{package}
\name{mvnimpute-package}
\alias{mvnimpute-package}
\title{mvnimpute: Implementation of multiple imputation involving the missing and censored values}
\description{
The mvnimpute package implements multiple imputation for concurrently handling missing and censored values based on the joint normal
model assumption.
}
\author{
Hesen Li
}
