% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnimpute.R
\docType{package}
\name{mvnimpute}
\alias{mvnimpute}
\title{mvnimpute: Implementation of multiple imputation involving the missing and censored values}
\description{
The mvnimpute package implements multiple imputation involving the missing and censored values
}
\author{
Hesen Li
}
