% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes.dat}
\alias{nhanes.dat}
\title{Combined NHANES dataset from release cycle 2011 - 2016}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://www.cdc.gov/nchs/nhanes/index.htm}
}
\usage{
nhanes.dat
}
\description{
Combined NHANES dataset from release cycle 2011 - 2016
}
\keyword{datasets}
