% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.calc.R
\name{param.calc}
\alias{param.calc}
\title{CC and AC parameters calculation function}
\usage{
param.calc(data)
}
\arguments{
\item{data}{data set containing the missing and censored values}
}
\value{
A list of four containing the available-case mean for each variable,
the complete-case covariance matrix, the available-case mean and variance for each variable.
}
\description{
Calculates the complete-case and available-case parameters
}
