% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_calc}
\alias{param_calc}
\title{Calculate the CC and AC parameters}
\usage{
param_calc(data)
}
\arguments{
\item{data}{a list object including the two matrices for the lower and upper bounds of the data.}
}
\description{
This function calculates the complete cases (CC) and available cases (AC) mean and variance values,
excluding the missing and censored values.
}
