% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_imputation.R
\name{sim.func}
\alias{sim.func}
\title{Multiple imputation function}
\usage{
sim.func(
  iter,
  prior.param,
  ini.vals,
  dat,
  miss.indx,
  miss.pos,
  censor.indx,
  censor.pos,
  censor.val,
  censor.type = "interval"
)
}
\arguments{
\item{iter}{Iteration number}

\item{prior.param}{List of prior parameter values}

\item{ini.vals}{List of initial values}

\item{dat}{Dataset to do multiple imputation}

\item{miss.indx}{Matrix of missing data index}

\item{miss.pos}{Position of variables with missing values in the original dataset}

\item{censor.indx}{Matrix of censoring data index}

\item{censor.pos}{Position of variables with censored values in the orignal dataset}

\item{censor.val}{List containing cutoff values for the censored variables}

\item{censor.type}{Logical variable specifying the censoring type}
}
\description{
This function implements the multiple imputation of the missing and censored values
}
