% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated.dat}
\alias{simulated.dat}
\title{Simulated continuous data with missing and censored values}
\format{
A list including data matrix with 200 rows and 4 variables and associated indicator matrix:
\describe{
   \item{y}{Outcome variable to be used in the regression model after imputation}
   \item{x1}{First covariate variable subject to MAR missing and non-informative censored values}
   \item{x2}{Second covariate variable subject to MAR missing and non-informative censored values}
   \item{x3}{Third covariate variable that is fully observed}
}
}
\usage{
simulated.dat
}
\description{
A dataset including simulated data with missing and censored values
}
\details{
A simulated dataset and its associated indicator matrix are included into a list. In the
indicator matrix, 0 stands for the missing values, 1 stands for the observed values, and 3 stands
for the left censored values.
}
\keyword{datasets}
