% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.imputation.R
\name{single.imputation}
\alias{single.imputation}
\title{Initial missing data fill-in function}
\usage{
single.imputation(data)
}
\arguments{
\item{data}{a list of data containing the information of the missing and censored values}
}
\value{
a complete dataset with missing and censored values filled in.
}
\description{
Fills in data with missing and censored values for multiple imputation
}
\details{
The unobserved values are initially filled in by some single imputation methods.
Currently, the missing values are generated from the normal distribution with complete-case
mean and variance as the mean and variance, and the censored values are generated from the
truncated normal distribution with complete-case parameters.
}
