% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulization_plot.R
\name{visual.plot}
\alias{visual.plot}
\title{Function for missing and censored data visualization}
\usage{
visual.plot(data, miss.index = NULL, censor.index = NULL, title = NULL)
}
\arguments{
\item{data}{dataset that contains the missing and censored values}

\item{miss.index}{missing index indicating the missing data in the original dataset}

\item{censor.index}{censoring index indicating the censored data in the original dataset}

\item{title}{the title of the plot. Default is NULL.}
}
\value{
graph that shows the details of the different type of data in the dataset
}
\description{
Draws the plots that graphically presents the percentage of the missing, censored and observed data
}
\details{
The function draws the plot that graphically shows the percentage of the missing, censored and observed
data in the dataset. Column names of the \code{miss.indx} and \code{censor.indx} matrices should correspond
to the respective variable names that are subject to either missing or censoring, and those names should be
handled beforehand.
}
