% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_plot.R
\name{visualization.plot}
\alias{visualization.plot}
\title{Function for missing and censored data visualization}
\usage{
visualization.plot(data.ind, title = NULL)
}
\arguments{
\item{data.ind}{dataset with missing and censored values}

\item{title}{a}
}
\value{
The plot that shows the details of the different type of data in the dataset
}
\description{
Draws the plots that graphically presents the percentage of the missing, censored and observed data
}
\details{
The function draws the plot that graphically shows the percentage of the missing, censored and observed
data in the dataset. Column names of the \code{miss.indx} and \code{censor.indx} matrices should correspond
to the respective variable names that are subject to either missing or censoring, and those names should be
handled beforehand.
}
