% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plot.R
\name{avg.plot}
\alias{avg.plot}
\title{Average values plot function}
\usage{
avg.plot(
  dat,
  iter,
  x.lab = "Iteration number",
  y.lab = "Average of simulated values",
  title = NULL,
  details = FALSE
)
}
\arguments{
\item{dat}{dataset containing the simulated values.}

\item{iter}{number of iterations for running multiple imputation.}

\item{x.lab}{label of the x axis in the generated plot, default is set to "Iteration number".}

\item{y.lab}{label of the y axis in the generated plot, default is set to "Average of simulated values".}

\item{title}{title of the generated plot.}

\item{details}{logical variable to specify whether the average simulate values are returned, default is set to TRUE.}
}
\value{
The plot of averaged values across iterations. If \code{details} = TRUE. A matrix containing the averaged values of all the variables across iterations.
}
\description{
Generates average simulated values of the parameter values
}
\details{
This function calculates the averaged simulated values across simulations.
}
