% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{calcu.acf}
\alias{calcu.acf}
\title{Autocorrelation function}
\usage{
calcu.acf(dat, lag, plot = FALSE, title = NULL, details = TRUE)
}
\arguments{
\item{dat}{dataset to calculate the autocorrelation}

\item{lag}{lag of autocorrelation to calulate}

\item{plot}{logical variable to specify whether the plot is generated, default is set to FALSE}

\item{title}{title of the generate autocorrelation plot}

\item{details}{logical variable to specify whether the autocorrelation values are returned, default is set to TRUE}
}
\description{
This function calculates the autocorrelations of all the variables in the dataset
}
