% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calcu.param}
\alias{calcu.param}
\title{Function to calculate complete case and avaliable case parameters}
\usage{
calcu.param(dat)
}
\arguments{
\item{dat}{dataset to compute the complete case and available case parameters}
}
\description{
Function to calculate complete case and avaliable case parameters
}
