% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calcu.param}
\alias{calcu.param}
\title{CC and AC parameters calculation function}
\usage{
calcu.param(data)
}
\arguments{
\item{data}{dataset containing the variables with missing and censored values}
}
\value{
A list of length 5 containing the available-case mean and variance for each variable,
the complete-case covariance matrix, mean and variance for each variable.
}
\description{
Calculates the complete-case and available-case parameters
}
