% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.plot.R
\name{conv.plot}
\alias{conv.plot}
\title{Convergence plot function}
\usage{
conv.plot(
  data,
  start,
  end,
  x.lab = "Iteration number",
  y.lab = "Simulated values",
  title = NULL
)
}
\arguments{
\item{data}{dataset containing the simulated values}

\item{start}{the number of cycle to start}

\item{end}{the number of cycle to end}

\item{x.lab}{label of the x axis in the generated plot, default is set to "Iteration number"}

\item{y.lab}{label of the y axis in the generated plot, default is set to "Simulated values"}

\item{title}{title of the generated plot.}
}
\value{
The plot of simulated values across iterations.
}
\description{
Generates the convergence plots for the parameter values
}
\details{
The function generates the trace plot of simulated values across iterations.
\code{iter} can be any number of iterations you want to draw, the corresponding number of rows
of the data is \code{iter} + 1.
}
