% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{conv.plot}
\alias{conv.plot}
\title{Convergence plot function}
\usage{
conv.plot(
  dat,
  iter,
  x.lab = "Iteration Number",
  y.lab = "Simulated values",
  title = NULL
)
}
\arguments{
\item{dat}{The dataset containing the values to be plotted}

\item{iter}{Iteration number of the simulation}

\item{x.lab}{Label of the x axis in the generated plot, default is set to be "Iteration number"}

\item{y.lab}{Label of the y axis in the generated plot, default is set to be "Simulated values"}

\item{title}{Title in the convergence plot}
}
\value{
Convergence plot of your desired variable
}
\description{
This function generates the convergence plots for the parameter values
}
