% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{conv.plot}
\alias{conv.plot}
\title{Convergence plot function}
\usage{
conv.plot(
  dat,
  iter,
  xlab = "Iteration Number",
  ylab = "Simulated values",
  title = NULL
)
}
\arguments{
\item{dat}{The dataset containing the values to be plotted}

\item{iter}{Iteration number of the simulation}

\item{xlab}{Label of the x axis in the convergence plot, default is set to be "Iteration number"}

\item{ylab}{Label of the y axis in the convergence plot, default is set to be "Simulated values"}

\item{title}{Title in the convergence plot}
}
\value{
Convergence plot of your desired variable
}
\description{
Function to generate convergence plots
}
