% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{dat.gen}
\alias{dat.gen}
\title{Data generating function}
\usage{
dat.gen(n, p, m.vec, sig, miss.num, censor.num)
}
\arguments{
\item{n}{number of observations to be generated.}

\item{p}{number of variables to be generated.}

\item{m.vec}{specified mean vector.}

\item{sig}{specified covariance matrix.}

\item{miss.num}{number of variables having missing values.}

\item{censor.num}{number of variables having censored values.}
}
\value{
A list containing the full data, the complete data with missing and censoring information applied,
  the missing and censoring index, positions of the variables including the missing or censored values, or
  the fully observed variable, and the two limits of the censored values
}
\description{
Generates multivariate normal data.
}
\details{
The function generates the multivariate normal data that can be used to verify the correctness of the multiple imputation
algorithm. Users have to specify the mean vector, variance covariance matrix, sample size and dimension of the generated data, and
the desired number of variables subject to missing and censoring, respectively. Currently, it only supports generating the data with
a certain type of MAR missing mechanism and interval censoring mechanism, in which case there should be at least one variable that is fully
observed in the dataset.
}
\examples{
n <- 1000
p <- 3
m.vec <- c(3, 2, 1)
V <- diag(p)
miss.num <- 1
censor.num <- 1
dat <- dat.gen(n, m.vec, V, p, miss.num, censor.num)

}
