% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{dat.gen}
\alias{dat.gen}
\title{Data generating function}
\usage{
dat.gen(n, p, m.vec, sig, miss.num, censor.num)
}
\arguments{
\item{n}{number of observations to be generated}

\item{p}{number of variables to be generated}

\item{m.vec}{specified mean vector}

\item{sig}{specified covariance matrix}

\item{miss.num}{number of variables having missing values}

\item{censor.num}{number of variables having censored values}
}
\value{
A list of 10 including:

\code{full.dat}: The fully observed complete data before the missing and censoring information applied.

\code{comp.dat}: The incomplete data after the missing and censoring information applied.

\code{miss.indx}: The matrix containing the missing index for the missing data. 1 for missing values, 0 for oberved values.

\code{miss.pos}: The vector indicating the position of the missing variables in the original dataset.

\code{miss.point}: The vector containing the cutoff values for the missing data. For the observed data that is less than or equal to the cutoff value, the corresponding missing variables are set to be missing.

\code{censor.pos}: The vector indicating the position of the censored variables in the original dataset.

\code{censor.indx}: The matrix containing the censoring index for the censored data. 1 for censored values, 0 otherwise.

\code{C1}: The vector containing the lower limits of the censored values.

\code{C2}: The vector containing the upper limits of the censored values.

\code{full.pos}: The vector indicating the position of the fully observed variables.
}
\description{
Generates multivariate normal data
}
\details{
The function generates the multivariate normal data that can be used to verify the correctness of the multiple imputation
algorithm. Users have to specify the mean vector, variance covariance matrix, sample size and dimension of the generated data, and
the desired number of variables subject to missing and censoring, respectively. Currently, it only supports generating the data with
a certain type of MAR missing mechanism and interval censoring mechanism, in which case there should be at least one variable that is fully
observed in the dataset.
}
\examples{
n <- 1000
p <- 3
m.vec <- c(3, 2, 1)
V <- diag(p)
miss.num <- 1
censor.num <- 1
dat <- dat.gen(n, p, m.vec, V, miss.num, censor.num)

}
