% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{data.generation}
\alias{data.generation}
\title{Data generating function}
\usage{
data.generation(
  n,
  p,
  mu,
  Sigma,
  miss.pos,
  miss.percent,
  miss.type = "MCAR",
  censor.pos,
  censor.percent,
  censor.type = "interval"
)
}
\arguments{
\item{n}{number of observations to be generated}

\item{p}{number of variables to be generated}

\item{mu}{specified mean vector}

\item{Sigma}{specified covariance matrix}

\item{miss.pos}{a}

\item{miss.percent}{b}

\item{miss.type}{c}

\item{censor.pos}{d}

\item{censor.percent}{e}

\item{censor.type}{f}
}
\description{
Generates multivariate normal data
}
\details{
This function generates the multivariate normal data that can be used to
verify the correctness of the multiple imputation algorithm. Users have to specify
the mean vector, variance covariance matrix, sample size and dimension of the generated
data, and the desired number of variables subject to missing and censoring, respectively.
Currently, it only supports generating the data with a certain type of MAR missing mechanism
and interval censoring mechanism, in which case there should be at least one variable that is
fully observed in the dataset.
}
