% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impute.R
\name{fill.dat}
\alias{fill.dat}
\title{Initial data fill-in fucntion}
\usage{
fill.dat(dat, miss.pos, censor.pos, method = "normal")
}
\arguments{
\item{dat}{dataset with missing and censored values}

\item{miss.pos}{Position of variables with missing values in the original dataset}

\item{censor.pos}{Position of variables with censored values in the original dataset}

\item{method}{method to be used for initial filled-in values}
}
\value{
a complete dataset with missing values filled in
}
\description{
Fill in data with missing and censored values for multiple imputation
}
