% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impute.R
\name{initial.impute}
\alias{initial.impute}
\title{Initial missing data fill-in function}
\usage{
initial.impute(
  data,
  miss.index = NULL,
  miss.pos = NULL,
  censor.index = NULL,
  censor.pos = NULL
)
}
\arguments{
\item{data}{dataset with missing and censored values}

\item{miss.index}{matrix containing the missing index}

\item{miss.pos}{vector containing the positions of variables with missing values in the original dataset}

\item{censor.index}{matrix containing the censoring index}

\item{censor.pos}{vector containing the positions of variables with censored values in the original dataset}
}
\value{
a complete dataset with missing and censored values filled in
}
\description{
Fills in data with missing and censored values for multiple imputation
}
\details{
The unobserved values are initially filled in by some single imputation methods.
Currently, it only supports generating random values from the normal distribution with the
mean and variance as the complete-case mean and variance, respectively.
}
