% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{marg.plot}
\alias{marg.plot}
\title{Marginal density plots function}
\usage{
marg.plot(data, title = NULL)
}
\arguments{
\item{data}{dataset containing all the variables}

\item{title}{title of the generated plots}
}
\value{
Marginal density plot for each variable in the dataset
}
\description{
Draws the marginal density plots for each variable
}
