% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{marg.plot}
\alias{marg.plot}
\title{Marginal plots function}
\usage{
marg.plot(dat, title = NULL)
}
\arguments{
\item{dat}{Dataset containing all the variables}

\item{title}{Titles of the generated plots}
}
\value{
Marginal density plots for each variable in the dataset
}
\description{
Draws the marginal density plots for each variable
}
