% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.imputation.R
\name{multiple.imputation}
\alias{multiple.imputation}
\title{Multiple imputation function}
\usage{
multiple.imputation(data, prior.params, starting.values, iter, verbose = TRUE)
}
\arguments{
\item{data}{a list of data containing the information for the missing and censored values}

\item{prior.params}{list of prior parameter specifications}

\item{starting.values}{list of starting values}

\item{iter}{number of rounds for doing multiple imputation}

\item{verbose}{boolean variable indicating whether the running status is printed in the console. Default is set to TRUE}
}
\description{
Implements the multiple imputation for both missing and censored values
}
\references{
Tanner, M., & Wong, W. (1987). The Calculation of Posterior Distributions by Data Augmentation.
\emph{Journal of the American Statistical Association}, \bold{82(398)}, 528-540.
}
