% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.imputation.R
\name{multiple.imputation}
\alias{multiple.imputation}
\title{Multiple imputation function}
\usage{
multiple.imputation(data, prior.params, initial.values, iter, verbose = TRUE)
}
\arguments{
\item{data}{a list of data containing the lower and upper bounds information for the missing and censored values.}

\item{prior.params}{list of prior parameter specifications.}

\item{initial.values}{list of initial values.}

\item{iter}{number of rounds for doing multiple imputation.}

\item{verbose}{boolean variable indicating whether the running status is printed in the console. Default is set to TRUE.}
}
\description{
This function implements the multiple imputation for multivariate data including both the missing and censored values.
The implemented multiple imputation algorithm is based on the data augmentation algorithm proposed by Tanner and Wong (1987).
The Gibbs sampling algorithm is adopted to update the model parameters and draw imputations of the coarse data.
}
\references{
Tanner, M., & Wong, W. (1987). The Calculation of Posterior Distributions by Data Augmentation.
\emph{Journal of the American Statistical Association}, \bold{82(398)}, 528-540.
}
