% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnimpute-package.R
\docType{package}
\name{mvnimpute-package}
\alias{mvnimpute-package}
\title{mvnimpute: Multiple imputation for multivariate data with missing and censored values}
\description{
The mvnimpute package implements multiple imputation for simultaneously imputing missing and censored values based on the joint normal
model assumption.
}
\author{
Hesen Li
}
