% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple imputation.R
\name{sim.func}
\alias{sim.func}
\title{Multiple imputation function}
\usage{
sim.func(
  iter,
  prior.param,
  ini.vals,
  dat,
  censoring.val,
  missing.indx = NULL,
  censoring.indx = NULL,
  censoring.type = "left"
)
}
\arguments{
\item{iter}{Number of iterations}

\item{prior.param}{A list containing the parameter values of the prior distribution}

\item{ini.vals}{A list containing the initial values}

\item{dat}{Dataset to do multiple imputation}

\item{censoring.val}{A matrix containing the censoring information}

\item{missing.indx}{A matrix containing the missing indices}

\item{censoring.indx}{A matrix containing the censoring indices}

\item{censoring.type}{Categorical variable specifying whether the data is right censored, left censored, or interval censored}
}
\description{
Function to do multiple imputations
}
