% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_imputation.R
\name{single.imputation}
\alias{single.imputation}
\title{Initial missing data fill-in function}
\usage{
single.imputation(data)
}
\arguments{
\item{data}{dataset with missing and censored values}
}
\value{
a complete dataset with missing and censored values filled in
}
\description{
Fills in data with missing and censored values for multiple imputation
}
\details{
The unobserved values are initially filled in by some single imputation methods.
Currently, it only supports generating random values from the normal distribution with the
mean and variance as the complete-case mean and variance, respectively.
}
